Moved some stuff around (like LookAt to Transform instead of Camera). Recompiled the SFML libraries with Visual C++ 2008. The old versions were giving some warnings. Cleaned up some header files with more comments.

+++ inc/vapor/math/Matrix4x4.h
@@ -35,8 +35,8 @@ public:
 	static Matrix4x4 createPerspectiveProjection( float fov, float ar, float near_, float far_ );
 
 	// Gets an orthographic matrix from a set of parameters.
-	static Matrix4x4 createOrthographicProjection( double left, double right, double bottom, double top,
-		double near_, double far_);
+	static Matrix4x4 createOrthographicProjection( float left, float right, float bottom, float top,
+		float near_, float far_);
 
 	// Identity matrix.
 	static const Matrix4x4 Identity;
